requirements.txt
alembic>=1.17.2
chromadb>=1.3.5
groq>=0.36.0
langchain>=1.0.8
langchain-community>=0.4.1
langchain-groq>=1.0.1
passlib[argon2]>=1.7.4
psycopg2-binary>=2.9.9
pypdf2>=3.0.1
pymupdf>=1.24.10
pytesseract>=0.3.13
Pillow>=11.0.0
python-docx>=1.2.0
sqlmodel>=0.0.31
streamlit>=1.51.0
tiktoken>=0.12.0
